Remove dependency on GetScsiUvmPath from WCOW-isolated mounts
The WCOW-isolated SCSI mount process currently works as follows:
- In resources_wcow.go, go through each mount on the OCI spec, and if it
  is a SCSI mount, add a mount to the UVM for it.
- in hcsdoc_wcow.go, go through each mount on the OCI spec, use
  GetScsiUvmPath to determine the guest path it was mounted to, and add
  an entry to the container doc for it.

This is quite hacky, as it relies on a 1:1 mapping between host VHDs and
mounts in the guest, and also because it requires us to re-query
information we've already been given. The SCSIMount object returned when
we mounted to the guest can already tell us the guest path.

This change resolves this problem by instead determing the set of guest
mounts that should be added to the container doc at the time when the
SCSI mounts are done, and saving it in the creation options. Then, when
we construct the actual container doc, we just grab those mounts and add
them in.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>