scsi: Remove VMAccessType, move to clients
Removes the VMAccessType parameter passed to AddSCSI* calls in the shim.
There was no real reason for the scsi code to handle this, as it's just
needed to set an ACL on the VHD to ensure the VMWP can access it, and
doesn't really have anything to do with scsi specifically.

Now it is each client's responsibility to set the appropriate ACL no the
VHD before calling the AddSCSI* functions.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>